requirements.txt
fastapi
uvicorn
psycopg2-binary
python-jose[cryptography]
bcrypt
python-multipart
python-dotenv
